Changes to accept optional input
file() was causing an error if an empty string was present in the config file.
Now we are using an empty list as input if the parameter is not provided in the config file (to allow for optional input).

Implementation of an error message if BAMs are provided without a reference genome (as nextflow default error is not explicit)

Changed comments in the config files to reflect the changes.

Remaining tabs were converted to 4 spaces in the code

Added correct versionning in confs